no commit message

+++ Components/BP.WF/Template/FrmNode.cs
@@ -289,7 +289,7 @@ namespace BP.WF.Template
             }
             set
             {
-                this.SetValByKey(FrmNodeAttr.IsEnableFWC, value);
+                this.SetValByKey(FrmNodeAttr.IsEnableFWC, (int)value);
             }
         }
         /// <summary>
